+++ examples/shop_app_example/lib/ui/ui.dart
@@ -1,0 +1,3 @@
+export 'initializer.dart';
+export 'router.dart';
+export 'theme.dart';