+++ packages/talker/lib/src/talker.dart
@@ -103,10 +103,9 @@ class Talker implements TalkerInterface {
   void handle(
     Object exception, [
     StackTrace? stackTrace,
-    String? msg,
-    // ErrorLevel? errorLevel,
+    dynamic msg,
   ]) {
-    final data = _errorHandler.handle(exception, stackTrace, msg);
+    final data = _errorHandler.handle(exception, stackTrace, msg?.toString());
     if (data is TalkerError) {
       _observersManager?.onError(data);
       _handleErrorData(data);
@@ -127,13 +126,12 @@ class Talker implements TalkerInterface {
   void handleError(
     Error error, [
     StackTrace? stackTrace,
-    String? msg,
-    // ErrorLevel? errorLevel,
+    dynamic msg,
   ]) {
     final data = TalkerError(
       error,
       stackTrace: stackTrace,
-      message: msg,
+      message: msg?.toString(),
       logLevel: LogLevel.error,
     );
     _handleErrorData(data);
@@ -147,13 +145,13 @@ class Talker implements TalkerInterface {
   void handleException(
     Exception exception, [
     StackTrace? stackTrace,
-    String? msg,
+    dynamic msg,
     // ErrorLevel? errorLevel,
   ]) {
     final data = TalkerException(
       exception,
       stackTrace: stackTrace,
-      message: msg,
+      message: msg?.toString(),
       logLevel: LogLevel.error,
     );
     _handleErrorData(data);
@@ -165,7 +163,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_log}
   @override
   void log(
-    String message, {
+    dynamic message, {
     LogLevel logLevel = LogLevel.debug,
     Object? exception,
     StackTrace? stackTrace,
@@ -183,7 +181,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_critical_log}
   @override
   void critical(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -193,7 +191,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_debug_log}
   @override
   void debug(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -203,7 +201,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_error_log}
   @override
   void error(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -213,7 +211,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_fine_log}
   @override
   void fine(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -223,7 +221,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_good_log}
   @override
   void good(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -233,7 +231,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_info_log}
   @override
   void info(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -243,7 +241,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_verbose_log}
   @override
   void verbose(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -253,7 +251,7 @@ class Talker implements TalkerInterface {
   /// {@macro talker_warning_log}
   @override
   void warning(
-    String msg, [
+    dynamic msg, [
     Object? exception,
     StackTrace? stackTrace,
   ]) {
@@ -281,7 +279,7 @@ class Talker implements TalkerInterface {
   }
 
   void _handleLog(
-    String message,
+    dynamic message,
     Object? exception,
     StackTrace? stackTrace,
     LogLevel logLevel, {
@@ -294,7 +292,7 @@ class Talker implements TalkerInterface {
       return;
     }
 
-    data = TalkerLog(message, logLevel: logLevel);
+    data = TalkerLog(message?.toString() ?? '', logLevel: logLevel);
     _handleLogData(data as TalkerLog, pen: pen);
   }