+++ packages/talker_logger/lib/src/talker_logger.dart
@@ -20,7 +20,7 @@ class TalkerLogger implements TalkerLoggerInterface {
 
   /// {@macro talker_logger_log}
   @override
-  void log(String msg, {LogLevel? level, AnsiPen? pen}) {
+  void log(dynamic msg, {LogLevel? level, AnsiPen? pen}) {
     final selectedPen = pen ?? settings.colors[level] ?? level.consoleColor;
     final selectedLevel = level ?? LogLevel.debug;
     if (_filter.shouldLog(msg, selectedLevel)) {
@@ -34,35 +34,35 @@ class TalkerLogger implements TalkerLoggerInterface {
 
   /// {@macro talker_logger_critical_log}
   @override
-  void critical(String msg) => log(msg, level: LogLevel.critical);
+  void critical(dynamic msg) => log(msg, level: LogLevel.critical);
 
   /// {@macro talker_logger_debug_log}
   @override
-  void debug(String msg) => log(msg);
+  void debug(dynamic msg) => log(msg);
 
   /// {@macro talker_logger_error_log}
   @override
-  void error(String msg) => log(msg, level: LogLevel.error);
+  void error(dynamic msg) => log(msg, level: LogLevel.error);
 
   /// {@macro talker_logger_fine_log}
   @override
-  void fine(String msg) => log(msg, level: LogLevel.fine);
+  void fine(dynamic msg) => log(msg, level: LogLevel.fine);
 
   /// {@macro talker_logger_good_log}
   @override
-  void good(String msg) => log(msg, level: LogLevel.good);
+  void good(dynamic msg) => log(msg, level: LogLevel.good);
 
   /// {@macro talker_logger_info_log}
   @override
-  void info(String msg) => log(msg, level: LogLevel.info);
+  void info(dynamic msg) => log(msg, level: LogLevel.info);
 
   /// {@macro talker_logger_verbose_log}
   @override
-  void verbose(String msg) => log(msg, level: LogLevel.verbose);
+  void verbose(dynamic msg) => log(msg, level: LogLevel.verbose);
 
   /// {@macro talker_logger_warning_log}
   @override
-  void warning(String msg) => log(msg, level: LogLevel.warning);
+  void warning(dynamic msg) => log(msg, level: LogLevel.warning);
 
   TalkerLogger copyWith({
     TalkerLoggerSettings? settings,