+++ packages/talker_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+First stable release! 🎉
+
+## 0.12.0
+  - **INFO** Update talker version to 0.12.0
+
 ## 0.11.0
   - **INFO** Update talker version to 0.11.0